+++ backend/mora/service/employee.py
@@ -352,13 +352,16 @@ def get_employee(id):
 
 
 @blueprint.route('/e/<uuid:employee_uuid>/terminate', methods=['POST'])
-@util.restrictargs()
+@util.restrictargs('validate')
 def terminate_employee(employee_uuid):
     """Terminates an employee and all of his roles beginning at a
     specified date. Except for the manager roles, which we vacate
     instead.
 
     .. :quickref: Employee; Terminate
+
+    :query boolean validate: Whether to apply validations to input;
+                             default is ``True``.
 
     :statuscode 200: The termination succeeded.
 
@@ -471,11 +474,14 @@ def get_employee_history(employee_uuid):
 
 
 @blueprint.route('/e/create', methods=['POST'])
-@util.restrictargs()
+@util.restrictargs('validate')
 def create_employee():
     """Create a new employee
 
     .. :quickref: Employee; Create
+
+    :query boolean validate: Whether to apply validations to input;
+                             default is ``True``.
 
     :statuscode 200: Creation succeeded.