+++ backend/mora/service/orgunit.py
@@ -823,11 +823,14 @@ def list_orgunit_tree(orgid):
 
 
 @blueprint.route('/ou/create', methods=['POST'])
-@util.restrictargs()
+@util.restrictargs('validate')
 def create_org_unit():
     """Creates new organisational unit
 
     .. :quickref: Unit; Create
+
+    :query boolean validate: Whether to apply validations to input;
+                             default is ``True``.
 
     :statuscode 200: Creation succeeded.
 
@@ -890,11 +893,14 @@ def create_org_unit():
 
 
 @blueprint.route('/ou/<uuid:unitid>/terminate', methods=['POST'])
-@util.restrictargs()
+@util.restrictargs('validate')
 def terminate_org_unit(unitid):
     """Terminates an organisational unit from a specified date.
 
     .. :quickref: Unit; Terminate
+
+    :query boolean validate: Whether to apply validations to input;
+                             default is ``True``.
 
     :statuscode 200: The termination succeeded.
     :statuscode 404: No such unit found.